[#][BE] fix: dog 수정, 삭제 시 이미지 없을 때 에러 수정

+++ backend/src/main/java/com/mtld/backend/service/dog/DogServiceImpl.java
@@ -127,7 +127,9 @@ public class DogServiceImpl implements DogService {
             return dog.getId();
         }
         // 기존 이미지 삭제
-        amazonS3Client.deleteObject(bucket, dog.getFileName());
+        if (dog.getFileName() != null) {
+            amazonS3Client.deleteObject(bucket, dog.getFileName());
+        }
 
         String originName = image.getOriginalFilename();
         String fileName = createFileName(originName);
@@ -158,9 +160,10 @@ public class DogServiceImpl implements DogService {
         if (!dog.getUser().equals(user)) {
             throw new AuthException("권한이 없습니다.");
         }
-
         // 기존 이미지 삭제
-        amazonS3Client.deleteObject(bucket, dog.getFileName());
+        if (dog.getFileName() != null) {
+            amazonS3Client.deleteObject(bucket, dog.getFileName());
+        }
 
         dogRepository.delete(dog);
     }
